requirements.txt
altgraph==0.17.3
anyio==3.7.1
camelot-py==0.11.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
copier==8.0.0
cryptography==41.0.1
decorator==5.1.1
dunamai==1.17.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
flet==0.8.1
flet-core==0.8.1
flet-runtime==0.8.1
funcy==2.0
ghostscript==0.7
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
macholib==1.16.2
MarkupSafe==2.1.3
numpy==1.25.1
oauthlib==3.2.2
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
pdfminer.six==20221105
plumbum==1.8.2
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
pypdf==3.12.1
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==1.10.0
repath==0.9.0
six==1.16.0
sniffio==1.3.0
tabulate==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
